Extract toScopeState out as a util function
